Move more brain handlers into subdirs
