+++ src/brain/gocd/gocdNoDeploysAlert/index.test.ts
@@ -4,7 +4,7 @@ import payloadRaw from '@test/payloads/gocd/gocd-stage-building.json';
 import { MockedBolt } from '@test/utils/testTypes';
 
 import * as slackblocks from '@/blocks/slackBlocks';
-import { DB_TABLE_STAGES } from '@/brain/saveGoCDStageEvents';
+import { DB_TABLE_STAGES } from '@/brain/gocd/saveGoCDStageEvents';
 import { buildServer } from '@/buildServer';
 import {
   FEED_DEV_INFRA_CHANNEL_ID,
