+++ src/utils/db/getLatestDeploy.test.ts
@@ -1,4 +1,4 @@
-import { DB_TABLE_STAGES } from '@/brain/saveGoCDStageEvents';
+import { DB_TABLE_STAGES } from '@/brain/gocd/saveGoCDStageEvents';
 import { db } from '@utils/db';
 
 import { FINAL_STAGE_NAMES } from '../gocd/gocdHelpers';
